Reorganization
time Module --> dynamic library

+++ scilab/modules/time/sci_gateway/gw_time.c
@@ -4,6 +4,7 @@
 /* Allan CORNET */
 /*-----------------------------------------------------------------------------------*/
 #include "MALLOC.h"
+#include "sciprint.h"
 #if _MSC_VER
 #include <Windows.h>
 extern char *GetExceptionString(DWORD ExceptionCode);
@@ -36,7 +37,7 @@ int C2F(gw_time)()
 		#ifndef _DEBUG
 		_try
 		{
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 		}
 		_except (EXCEPTION_EXECUTE_HANDLER)
 		{
@@ -45,10 +46,10 @@ int C2F(gw_time)()
 			if (ExceptionString) {FREE(ExceptionString);ExceptionString=NULL;}
 		}
 		#else
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 		#endif
 	#else
-		(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+		(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 	#endif
 
 	return 0;
